backend/requirements.txt
annotated-doc==0.0.4
anthropic>=0.39.0
annotated-types==0.7.0
anyio==4.12.1
boto3==1.42.48
botocore==1.42.48
click==8.3.1
fastapi==0.129.0
h11==0.16.0
idna==3.11
jmespath==1.1.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
s3transfer==0.16.0
six==1.17.0
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pyjwt>=2.8.0
bcrypt>=4.0.0
cryptography>=41.0.0
